Move implementations into lib.rs
Also, we don't need extren crate anymore, it's gone since Rust 2018 :)
