+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "split_benches"
+name = "splitter"
 version = "0.1.0"
 authors = ["Sam Schlinkert <sschlinkert@gmail.com>"]
 edition = "2018"
